requirements.txt
streamlit==1.18.1
torch==1.13.1
transformers==4.24.0
fastapi==0.95.1
langchain==0.0.143
Pillow==9.3.0
openai==0.27.0
requests==2.28.1
gradio==3.10.1
opencv-python==4.6.0.66
tesseract==0.1.3
PyPDF2==1.26.0
faiss-cpu==1.7.3
chromadb==0.4.2
